Move extractors one directory up
