+++ YoutubeExplode/Bridge/ChannelPageExtractor.cs
@@ -3,7 +3,7 @@ using AngleSharp.Html.Dom;
 using YoutubeExplode.Utils;
 using YoutubeExplode.Utils.Extensions;
 
-namespace YoutubeExplode.Bridge.Extractors;
+namespace YoutubeExplode.Bridge;
 
 internal partial class ChannelPageExtractor
 {
