+++ YoutubeExplode/Bridge/ClosedCaptionExtractor.cs
@@ -5,7 +5,7 @@ using System.Xml.Linq;
 using YoutubeExplode.Utils;
 using YoutubeExplode.Utils.Extensions;
 
-namespace YoutubeExplode.Bridge.Extractors;
+namespace YoutubeExplode.Bridge;
 
 internal class ClosedCaptionExtractor
 {
