+++ YoutubeExplode/Bridge/ClosedCaptionPartExtractor.cs
@@ -3,7 +3,7 @@ using System.Xml.Linq;
 using YoutubeExplode.Utils;
 using YoutubeExplode.Utils.Extensions;
 
-namespace YoutubeExplode.Bridge.Extractors;
+namespace YoutubeExplode.Bridge;
 
 internal class ClosedCaptionPartExtractor
 {
