+++ YoutubeExplode/Bridge/PlayerClosedCaptionTrackInfoExtractor.cs
@@ -4,7 +4,7 @@ using System.Text.Json;
 using YoutubeExplode.Utils;
 using YoutubeExplode.Utils.Extensions;
 
-namespace YoutubeExplode.Bridge.Extractors;
+namespace YoutubeExplode.Bridge;
 
 internal class PlayerClosedCaptionTrackInfoExtractor
 {
