+++ YoutubeExplode/Bridge/PlayerResponseExtractor.cs
@@ -7,7 +7,7 @@ using System.Text.RegularExpressions;
 using YoutubeExplode.Utils;
 using YoutubeExplode.Utils.Extensions;
 
-namespace YoutubeExplode.Bridge.Extractors;
+namespace YoutubeExplode.Bridge;
 
 internal partial class PlayerResponseExtractor
 {
