+++ YoutubeExplode/Bridge/PlayerStreamInfoExtractor.cs
@@ -5,7 +5,7 @@ using System.Text.RegularExpressions;
 using YoutubeExplode.Utils;
 using YoutubeExplode.Utils.Extensions;
 
-namespace YoutubeExplode.Bridge.Extractors;
+namespace YoutubeExplode.Bridge;
 
 internal class PlayerStreamInfoExtractor : IStreamInfoExtractor
 {
