+++ YoutubeExplode/Bridge/PlaylistVideoExtractor.cs
@@ -5,16 +5,16 @@ using System.Text.Json;
 using YoutubeExplode.Utils;
 using YoutubeExplode.Utils.Extensions;
 
-namespace YoutubeExplode.Bridge.Extractors;
+namespace YoutubeExplode.Bridge;
 
-internal class SearchResultVideoExtractor
+internal class PlaylistVideoExtractor
 {
     private static readonly string[] DurationFormats = {@"m\:ss", @"mm\:ss", @"h\:mm\:ss", @"hh\:mm\:ss"};
 
     private readonly JsonElement _content;
     private readonly Memo _memo = new();
 
-    public SearchResultVideoExtractor(JsonElement content) => _content = content;
+    public PlaylistVideoExtractor(JsonElement content) => _content = content;
 
     public string? TryGetVideoId() => _memo.Wrap(() =>
         _content
@@ -37,7 +37,7 @@ internal class SearchResultVideoExtractor
             .ConcatToString()
     );
 
-    private JsonElement? TryGetVideoAuthorDetails() => _memo.Wrap(() =>
+    private JsonElement? TryGetAuthorDetails() => _memo.Wrap(() =>
         _content
             .GetPropertyOrNull("longBylineText")?
             .GetPropertyOrNull("runs")?
@@ -52,13 +52,13 @@ internal class SearchResultVideoExtractor
     );
 
     public string? TryGetVideoAuthor() => _memo.Wrap(() =>
-        TryGetVideoAuthorDetails()?
+        TryGetAuthorDetails()?
             .GetPropertyOrNull("text")?
             .GetStringOrNull()
     );
 
     public string? TryGetVideoChannelId() => _memo.Wrap(() =>
-        TryGetVideoAuthorDetails()?
+        TryGetAuthorDetails()?
             .GetPropertyOrNull("navigationEndpoint")?
             .GetPropertyOrNull("browseEndpoint")?
             .GetPropertyOrNull("browseId")?
@@ -66,6 +66,12 @@ internal class SearchResultVideoExtractor
     );
 
     public TimeSpan? TryGetVideoDuration() => _memo.Wrap(() =>
+        _content
+            .GetPropertyOrNull("lengthSeconds")?
+            .GetStringOrNull()?
+            .ParseDoubleOrNull()?
+            .Pipe(TimeSpan.FromSeconds) ??
+
         _content
             .GetPropertyOrNull("lengthText")?
             .GetPropertyOrNull("simpleText")?
