+++ YoutubeExplode/Bridge/SearchResultsExtractor.cs
@@ -5,7 +5,7 @@ using System.Text.Json;
 using YoutubeExplode.Utils;
 using YoutubeExplode.Utils.Extensions;
 
-namespace YoutubeExplode.Bridge.Extractors;
+namespace YoutubeExplode.Bridge;
 
 internal partial class SearchResultsExtractor
 {
