+++ YoutubeExplode/Bridge/ThumbnailExtractor.cs
@@ -2,7 +2,7 @@
 using YoutubeExplode.Utils;
 using YoutubeExplode.Utils.Extensions;
 
-namespace YoutubeExplode.Bridge.Extractors;
+namespace YoutubeExplode.Bridge;
 
 internal class ThumbnailExtractor
 {
