+++ YoutubeExplode/Bridge/VideoWatchPageExtractor.cs
@@ -6,7 +6,7 @@ using AngleSharp.Html.Dom;
 using YoutubeExplode.Utils;
 using YoutubeExplode.Utils.Extensions;
 
-namespace YoutubeExplode.Bridge.Extractors;
+namespace YoutubeExplode.Bridge;
 
 internal partial class VideoWatchPageExtractor
 {
