+++ YoutubeExplode/Channels/ChannelController.cs
@@ -2,7 +2,6 @@
 using System.Threading;
 using System.Threading.Tasks;
 using YoutubeExplode.Bridge;
-using YoutubeExplode.Bridge.Extractors;
 using YoutubeExplode.Exceptions;
 
 namespace YoutubeExplode.Channels;
