+++ YoutubeExplode/Videos/ClosedCaptions/ClosedCaptionController.cs
@@ -1,7 +1,7 @@
 ﻿using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
-using YoutubeExplode.Bridge.Extractors;
+using YoutubeExplode.Bridge;
 using YoutubeExplode.Utils;
 using YoutubeExplode.Utils.Extensions;
