+++ YoutubeExplode/Videos/Streams/StreamController.cs
@@ -1,7 +1,7 @@
 ﻿using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
-using YoutubeExplode.Bridge.Extractors;
+using YoutubeExplode.Bridge;
 
 namespace YoutubeExplode.Videos.Streams;
